♻️  macros: Refactor introspection macros
