+++ zlink-macros/src/introspect/mod.rs
@@ -1,3 +1,5 @@
 pub(crate) mod custom_type;
 pub(crate) mod reply_error;
 pub(crate) mod r#type;
+
+mod shared;
