+++ zlink-macros/src/introspect/type.rs
@@ -1,8 +1,10 @@
 use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
-use syn::{Data, DataEnum, DeriveInput, Error, Fields, FieldsNamed, FieldsUnnamed};
+use syn::{Data, DataEnum, DeriveInput, Error, Fields};
 
 use crate::utils;
+
+use super::shared;
 
 /// Main entry point for the Type derive macro.
 pub(crate) fn derive_type(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -69,89 +71,12 @@ fn generate_field_definitions(
     fields: &Fields,
     crate_path: &TokenStream2,
 ) -> Result<(Vec<TokenStream2>, Vec<TokenStream2>), Error> {
-    match fields {
-        Fields::Named(FieldsNamed { named, .. }) => {
-            let mut field_statics = Vec::new();
-            let mut field_refs = Vec::new();
-
-            for field in named {
-                let field_name = field
-                    .ident
-                    .as_ref()
-                    .ok_or_else(|| Error::new_spanned(field, "Field must have a name"))?;
-
-                let field_type = utils::remove_lifetimes_from_type(&field.ty);
-                let field_name_str = field_name.to_string();
-                let static_name =
-                    quote::format_ident!("FIELD_{}", field_name.to_string().to_uppercase());
-
-                let comments = utils::extract_doc_comments(&field.attrs);
-                let comment_objects: Vec<_> = comments
-                    .iter()
-                    .map(|c| quote! { &#crate_path::idl::Comment::new(#c) })
-                    .collect();
-
-                let field_static = quote! {
-                    static #static_name: #crate_path::idl::Field<'static> =
-                        #crate_path::idl::Field::new(
-                            #field_name_str,
-                            <#field_type as #crate_path::introspect::Type>::TYPE,
-                            &[#(#comment_objects),*]
-                        );
-                };
-
-                let field_ref = quote! { &#static_name };
-
-                field_statics.push(field_static);
-                field_refs.push(field_ref);
-            }
-
-            Ok((field_statics, field_refs))
-        }
-        Fields::Unnamed(FieldsUnnamed { unnamed, .. }) => Err(Error::new_spanned(
-            unnamed,
-            "Only named fields are supported",
-        )),
-        Fields::Unit => {
-            // Unit structs have no fields.
-            Ok((Vec::new(), Vec::new()))
-        }
-    }
+    shared::generate_field_definitions(fields, crate_path, None)
 }
 
 fn generate_enum_variant_definitions(
     data_enum: &DataEnum,
     crate_path: &TokenStream2,
 ) -> Result<Vec<TokenStream2>, Error> {
-    let mut variant_refs = Vec::new();
-
-    for variant in &data_enum.variants {
-        // Only support unit variants (no associated data).
-        match &variant.fields {
-            Fields::Unit => {
-                let variant_name = variant.ident.to_string();
-                let comments = utils::extract_doc_comments(&variant.attrs);
-                let comment_objects: Vec<_> = comments
-                    .iter()
-                    .map(|c| quote! { &#crate_path::idl::Comment::new(#c) })
-                    .collect();
-                let variant_ref = quote! { &#crate_path::idl::EnumVariant::new(#variant_name, &[#(#comment_objects),*]) };
-                variant_refs.push(variant_ref);
-            }
-            Fields::Named(_) => {
-                return Err(Error::new_spanned(
-                    variant,
-                    "Type derive macro only supports unit enum variants, not struct variants",
-                ));
-            }
-            Fields::Unnamed(_) => {
-                return Err(Error::new_spanned(
-                    variant,
-                    "Type derive macro only supports unit enum variants, not tuple variants",
-                ));
-            }
-        }
-    }
-
-    Ok(variant_refs)
+    shared::generate_enum_variant_definitions(data_enum, crate_path)
 }
